# backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Mozgoslav.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/_internal/devices/changed": {
      "post": {
        "tags": [
          "Mozgoslav.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioDeviceChangePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/_internal/hotkey/event": {
      "post": {
        "tags": [
          "Mozgoslav.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotkeyEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/dictation/push/{sessionId}": {
      "post": {
        "tags": [
          "Mozgoslav.Api"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushChunkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/dictation/{sessionId}/push": {
      "post": {
        "tags": [
          "Mozgoslav.Api"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AudioDeviceChangePayload": {
        "required": [
          "kind",
          "devices",
          "observedAt"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudioDeviceInfo"
            }
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AudioDeviceInfo": {
        "required": [
          "id",
          "name",
          "isDefault"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
      "HotkeyEvent": {
        "required": [
          "kind",
          "accelerator",
          "observedAt"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "accelerator": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PushChunkRequest": {
        "required": [
          "samples",
          "sampleRate",
          "offsetSeconds"
        ],
        "type": "object",
        "properties": {
          "samples": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          "sampleRate": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "offsetSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Mozgoslav.Api"
    }
  ]
}